pyproject.toml
[tool.poetry]
name = "spectra"
version = "0.0.1"
description = "An AI-powered companion that revolutionizes navigation for the visually impaired."
authors = ["Ayush <ayushgundawar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
opencv-python = "^4.8.1.78"
google-generativeai = "^0.2.2"
google-auth = "^2.23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
